PieceWiseLinearDiscontinuous is in 1 .h/.cc file
